Improves the listen method, adds thorough test
The listen method now takes an additional channel as a parameter,
which is a ticker channel. This is meant to send signals at regular
intervals, and push the tetromino down at some rate which is
determined by the level. If the tetromino can't move down any further,
then it triggers the NextTet. Everything here works perfectly with
unbuffered channels.