Removes channel handling from board controller
The board controller shouldn't be dealing with any notion of time, and
it was far too complex with multiple channels for simple things. This
alters the controller to drastically simplify it, and make modifying
it's state much easier.

The listen method is gone, and instead is a Tick method. This method
will atomically update the state of the controller, which is extremely
handy and much easier to test.

The idea is to push the higher level state into the Game, so things
just stay simple. This way when we deal with drawing code we know our
state transitions are atomic and we don't need to trace race
conditions.

As a result, our testing is also pretty simplified since it cuts out a
lot of the nonsense we dealt with just to setup tests.