Fixes rotation
Now all the rotations are unique. Adds testing to ensure this stays
true in the future. When we refer to pointers of values, we have to be
careful to get it right. This takes care of the problem be ensuring we
make proper allocations, and then also do the needed resassignment.